+++ RayTracer/src/interaction.h
@@ -3,6 +3,7 @@
 #include "ray.h"
 
 #include <vector>
+#include <memory>
 
 namespace rt 
 {
@@ -14,7 +15,7 @@ class Interaction
 public:
 	Point3 posn;
 	Vec3 normal;
-	double t = 0.0;
+	float t = 0.0f;
 	std::shared_ptr<Material> material;
 	bool front_face = false;
 
@@ -23,7 +24,7 @@ public:
 		/* Sets the interaction normal vector */
 		/* Note: the parameter `outward_normal` is assumed to have unit length */
 
-		front_face = Dot(ray.direction, outward_normal) < 0;
+		front_face = Dot(ray.direction, outward_normal) < 0.0f;
 		normal = front_face ? outward_normal : -outward_normal;
 	}
